requirements.txt
fastapi~=0.70.0
sqlmodel~=0.0.4
SQLAlchemy~=1.4.27
pytz~=2021.3
phonenumbers~=8.12.38
pydantic~=1.8.2
passlib~=1.7.4
click~=8.0.3
gunicorn~=20.1.0
uvicorn~=0.15.0
asyncio~=3.4.3
alembic~=1.7.5
python-jose~=3.3.0
fastapi_login~=1.7.3
asyncpg~=0.24.0
python-multipart~=0.0.5
databases~=0.5.3